JS-13-multi-module feat(*): add solution task
