+++ src/helper.js
@@ -1,0 +1,3 @@
+export const formatNumber = (number) => {
+  return number > 9 ? number : `0${number}`;
+}
